Changed the EventList to EventGrid
- Galen was using flexbox. It was a dumb decision
- Galen had to write JavaScript to calculate the width of flexbox
  given the number of items for each screen size
- Galen discovered display: grid and it is so simple and elegant
- It was 82 lines shorter (55.3% shorter) OMG Galen was so dumb
